refactor: Rename methods refactor: Restructure project folders
