+++ src/main/java/com/bovo/Bovo/modules/user/dto/request/SignupDto.java
@@ -1,4 +1,4 @@
-package com.bovo.Bovo.modules.user.dto;
+package com.bovo.Bovo.modules.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
