+++ src/main/java/com/bovo/Bovo/modules/user/repository/UserAuthRepositoryImpl.java
@@ -1,8 +1,6 @@
 package com.bovo.Bovo.modules.user.repository;
 
-import com.bovo.Bovo.modules.user.domain.Users;
-import com.bovo.Bovo.modules.user.domain.User_Auth;
-import com.bovo.Bovo.modules.user.domain.User_Auth;
+import com.bovo.Bovo.common.domain.User_Auth;
 import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
