+++ src/main/java/com/bovo/Bovo/modules/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.bovo.Bovo.modules.user.repository;
 
-import com.bovo.Bovo.modules.user.domain.Users;
+import com.bovo.Bovo.common.domain.Users;
 
 public interface UserRepository {
     Users save(Users user); // 회원가입: user를 DB에 저장
