+++ src/main/java/com/bovo/Bovo/modules/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.bovo.Bovo.modules.user.service;
 
-import com.bovo.Bovo.modules.user.dto.LoginDto;
-import com.bovo.Bovo.modules.user.dto.SignupDto;
+import com.bovo.Bovo.modules.user.dto.request.LoginDto;
+import com.bovo.Bovo.modules.user.dto.request.SignupDto;
 import jakarta.servlet.http.Cookie;
 
 public interface UserService {
